refactor: auth domain
