+++ backend/src/main/java/com/myocean/global/config/SwaggerConfig.java
@@ -1,4 +1,4 @@
-package com.myocean.config;
+package com.myocean.global.config;
 
 import io.swagger.v3.oas.models.Components;
 import io.swagger.v3.oas.models.OpenAPI;
